Two item view wrappers: checkable and non
So that setActivated() will be called on items that
are not checkable when some choice mode is enabled.
